rivos: Fixed freg_t freg(uint64_t f) to handle 32/64 bit FP values

+++ src/arch/riscv/regs/float.hh
@@ -99,7 +99,8 @@ static constexpr float64_t f64(freg_t r) { return r; }
 // Create generic floating point values from fixed size floats.
 static constexpr freg_t freg(float32_t f) { return {boxF32(f.v)}; }
 static constexpr freg_t freg(float64_t f) { return f; }
-static constexpr freg_t freg(uint_fast16_t f) { return {f}; }
+
+static constexpr freg_t freg(uint64_t f) { return {f}; }
 
 #define F32_SIGN ((uint32_t)1 << 31)
 #define F64_SIGN ((uint64_t)1 << 63)
